Fix lombok  issues

+++ backend/src/main/java/com/example/phoebe/dto/response/PublicNewsDto.java
@@ -3,7 +3,7 @@ package com.example.phoebe.dto.response;
 import java.time.ZonedDateTime;
 
 public class PublicNewsDto {
-    private Long id;
+    private Integer id;
     private String title;
     private String slug;
     private String teaser;
@@ -12,7 +12,7 @@ public class PublicNewsDto {
     // Constructors
     public PublicNewsDto() {}
 
-    public PublicNewsDto(Long id, String title, String slug, String teaser, ZonedDateTime publishedAt) {
+    public PublicNewsDto(Integer id, String title, String slug, String teaser, ZonedDateTime publishedAt) {
         this.id = id;
         this.title = title;
         this.slug = slug;
@@ -21,14 +21,14 @@ public class PublicNewsDto {
     }
 
     // Getters
-    public Long getId() { return id; }
+    public Integer getId() { return id; }
     public String getTitle() { return title; }
     public String getSlug() { return slug; }
     public String getTeaser() { return teaser; }
     public ZonedDateTime getPublishedAt() { return publishedAt; }
 
     // Setters
-    public void setId(Long id) { this.id = id; }
+    public void setId(Integer id) { this.id = id; }
     public void setTitle(String title) { this.title = title; }
     public void setSlug(String slug) { this.slug = slug; }
     public void setTeaser(String teaser) { this.teaser = teaser; }
